Flexible alignment of end labels
In some cases it would be nice to be able to control alignment of
the end labels. If labels are long they will stretch beyond the
borders of the sliders. This PR implements a way to avoid this by
introducing the `centerEndLabels` parameter. When this parameter is
false the first label will be left-aligned and the last label will
be right-aligned. Horizontal margin can be controlled by using the
new `horizontalMargin` parameter.